+++ pyriemann/classification.py
@@ -14,7 +14,7 @@ from sklearn.pipeline import make_pipeline
 from joblib import Parallel, delayed
 
 from .utils.kernel import kernel
-from .utils.mean import mean_covariance
+from .utils.mean import mean_covariance, mean_power
 from .utils.distance import distance
 from .tangentspace import FGDA, TangentSpace
 
@@ -715,3 +715,180 @@ class SVC(sklearnSVC):
         else:
             raise TypeError(f"kernel must be 'precomputed' or callable, is "
                             f"{self.kernel}.")
+
+
+class MeanField(BaseEstimator, ClassifierMixin, TransformerMixin):
+    """Classification by Minimum Distance to Mean Field.
+
+    Classification by Minimum Distance to Mean Field [1]_, using power means.
+
+    Parameters
+    ----------
+    power_list : list of float (default: [-1,0,+1])
+        Exponents of power means.
+    method_label : {'sum_means', 'inf_means'} (default: 'sum_means')
+        Method to combine labels:
+
+        * sum_means: it assigns the covariance to the class whom the sum of
+          distances to means of the field is the lowest;
+        * inf_means: it assigns the covariance to the class of the closest mean
+          of the field.
+    metric : string | dict (default: 'riemann')
+        The type of metric used for distance estimation during prediction.
+        See `distance` for the list of supported metric.
+
+    Attributes
+    ----------
+    covmeans_ : list of list
+        the class centroids.
+    classes_ : list
+        list of classes.
+
+    See Also
+    --------
+    MDM
+
+    Notes
+    -----
+    .. versionadded:: 0.2.8
+
+    References
+    ----------
+    .. [1] M Congedo, PLC Rodrigues, C Jutten. "The Riemannian Minimum Distance
+        to Means Field Classifier", BCI Conference 2019
+    """
+
+    def __init__(self, power_list=[-1, 0, 1], method_label='sum_means',
+                 metric='riemann', n_jobs=1):
+        """Init."""
+        self.power_list = power_list
+        self.method_label = method_label
+        self.metric = metric
+        self.n_jobs = n_jobs
+
+    def fit(self, X, y, sample_weight=None):
+        """Fit (estimates) the centroids.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+        y : ndarray, shape (n_matrices,)
+            Labels corresponding to each matrix.
+        sample_weight : None | ndarray shape (n_matrices,)
+            Weights of each matrix. If None, each matrix is treated with
+            equal weights.
+
+        Returns
+        -------
+        self : MeanField instance
+            The MeanField instance.
+        """
+        self.classes_ = np.unique(y)
+
+        if sample_weight is None:
+            sample_weight = np.ones(X.shape[0])
+
+        self.covmeans_ = {}
+        for p in self.power_list:
+            means_p = {}
+            for ll in self.classes_:
+                means_p[ll] = mean_power(
+                    X[y == ll],
+                    p,
+                    sample_weight=sample_weight[y == ll]
+                )
+            self.covmeans_[p] = means_p
+
+        return self
+
+    def _get_label(self, x, labs_unique):
+        m = np.zeros((len(self.power_list), len(labs_unique)))
+        for ip, p in enumerate(self.power_list):
+            for ill, ll in enumerate(labs_unique):
+                m[ip, ill] = distance(
+                    x, self.covmeans_[p][ll], metric=self.metric)**2
+
+        if self.method_label == 'sum_means':
+            ipmin = np.argmin(np.sum(m, axis=1))
+        elif self.method_label == 'inf_means':
+            ipmin = np.where(m == np.min(m))[0][0]
+        else:
+            raise TypeError('method_label must be sum_means or inf_means')
+
+        y = labs_unique[np.argmin(m[ipmin])]
+        return y
+
+    def predict(self, X):
+        """Get the predictions.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+
+        Returns
+        -------
+        pred : ndarray of int, shape (n_matrices,)
+            Predictions for each matrix according to the closest means field.
+        """
+        labs_unique = sorted(self.covmeans_[self.power_list[0]].keys())
+
+        pred = Parallel(n_jobs=self.n_jobs)(delayed(self._get_label)(
+            x, labs_unique)
+            for x in X)
+        return np.array(pred)
+
+    def _predict_distances(self, X):
+        """Helper to predict the distance. Equivalent to transform."""
+
+        dist = []
+        for x in X:
+            m = {}
+            for p in self.power_list:
+                m[p] = []
+                for ll in self.classes_:
+                    m[p].append(
+                        distance(
+                            x, self.covmeans_[p][ll], metric=self.metric
+                        )**2
+                    )
+            pmin = min(m.items(), key=lambda x: np.sum(x[1]))[0]
+            dist.append(np.array(m[pmin]))
+
+        return np.stack(dist)
+
+    def transform(self, X):
+        """Get the distance to each means field.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+
+        Returns
+        -------
+        dist : ndarray, shape (n_matrices, n_classes)
+            Distance to each means field according to the metric.
+        """
+        return self._predict_distances(X)
+
+    def fit_predict(self, X, y):
+        """Fit and predict in one function."""
+        self.fit(X, y)
+        return self.predict(X)
+
+    def predict_proba(self, X):
+        """Predict proba using softmax.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+
+        Returns
+        -------
+        prob : ndarray, shape (n_matrices, n_classes)
+            Probabilities for each class.
+        """
+        return softmax(-self._predict_distances(X)**2)